fix(findKeyboardDev): Workaround for macOS
WARNING: THIS CODE IS NOT MINE!!! THE DEVICE DETECTION ON MAC IS ENTIRELY VIDE-CODED ...
"go.bug.st/serial/enumerator requires cgo for MacOS builds which makes it hard to cross compile.
Vibe coded a workaround for the MacOS but it is entirely not tested and I ahve no clue how MacOS works!